Added `dependencies` module
* Added dependency on “server”
* Added “order” property in update format
