+++ Cargo.toml
@@ -10,3 +10,4 @@ path = "src/main.rs"
 piston_meta = "0.12.1"
 hyper = "0.6.9"
 range = "0.1.1"
+semver = "0.1.20"
