+++ src/update.rs
@@ -1,0 +1,1 @@
+//! Generate update information from dependency information.
